moved relationship_type to proper folder
