+++ src/aoptk/relationships/relationship_type.py
